Simplifies the csi related cmd packages
Removes that god-awful builder pattern and related tests.
- They provide no (actualy negative) value
- The related tests are just testing that we can write mocks and not
actual functionality
- They are totaly unreadable
- Fixing a test here takes longer than just simplifying the whole thing

Should be done for all other cmd packages, but didn't want to touch
everything
